refactor(shared): group backend support modules
