+++ cli/src/runtime.mjs
@@ -6,7 +6,7 @@ import {
   effectiveBackendPermissionMode,
   normalizeBackendProtocol,
   resolveBackendOwnership,
-} from '../../shared/backend-catalog.mjs'
+} from '../../shared/backend/catalog.mjs'
 import {
   loadRuntimeEnvironment,
   requireRealtimeFrontendConfiguration,
