+++ desktop/src/backend/detection-worker.mjs
@@ -1,7 +1,7 @@
 import { parentPort, workerData } from 'node:worker_threads'
 
-import { inspectBackendSetupsAsync } from '../../../shared/backend-setup.mjs'
-import { inspectBackendAuthentication } from '../../../shared/backend-auth-status.mjs'
+import { inspectBackendSetupsAsync } from '../../../shared/backend/setup.mjs'
+import { inspectBackendAuthentication } from '../../../shared/backend/auth-status.mjs'
 import { refreshProcessPath } from '../process-path.mjs'
 
 function compactComponent(component) {
