+++ desktop/src/backend/installer.mjs
@@ -1,11 +1,11 @@
 // 设置页"安装"按钮的安装编排：并发守卫、脚本确认、进度回调。
 // 与 main.mjs 的 IPC 接线分离，依赖全部注入，保证可单元测试；
-// 安装逻辑本身复用 shared/backend-install.mjs（与 CLI 同一份）。
+// 安装逻辑本身复用 shared/backend/install.mjs（与 CLI 同一份）。
 
 import {
   installBackend,
   installSupport,
-} from '../../../shared/backend-install.mjs'
+} from '../../../shared/backend/install.mjs'
 
 export function createBackendInstaller({
   env = process.env,
