+++ desktop/src/backend/management.mjs
@@ -1,8 +1,8 @@
 import { existsSync, readFileSync } from 'node:fs'
 import { parseEnv } from 'node:util'
 import { mergeSearchPath } from '../../../shared/path-environment.mjs'
-import { withBackendLifecycle } from '../../../shared/backend-install.mjs'
-import { backendDefinition } from '../../../shared/backend-catalog.mjs'
+import { withBackendLifecycle } from '../../../shared/backend/install.mjs'
+import { backendDefinition } from '../../../shared/backend/catalog.mjs'
 import { detectBackendSetups } from './detection.mjs'
 import { createBackendInstaller } from './installer.mjs'
 import { openBackendConfiguration } from './configuration.mjs'
