+++ desktop/src/settings-config.mjs
@@ -3,7 +3,7 @@ import {
   backendDefinition,
   normalizeBackendProtocol,
   resolveBackendOwnership,
-} from '../../shared/backend-catalog.mjs'
+} from '../../shared/backend/catalog.mjs'
 import {
   normalizeOrbSkinId,
   resolveOrbSkinId,
